refactor: Admin 페이지 내 컴포넌트 엔티티 중심으로 디렉토리 분리
