+++ src/refactoring/components/admin/coupon/CouponItem.tsx
@@ -1,4 +1,4 @@
-import { Coupon } from "../../../types";
+import { Coupon } from "../../../../types";
 
 interface CouponItemProps {
   coupon: Coupon;
