+++ src/refactoring/components/admin/coupon/CouponItemList.tsx
@@ -1,4 +1,4 @@
-import { Coupon } from "../../../types";
+import { Coupon } from "../../../../types";
 import { CouponItem } from "./CouponItem";
 
 interface CouponItemListProps {
