+++ src/refactoring/components/admin/coupon/index.ts
@@ -1,0 +1,1 @@
+export * from "./CouponManagementSection";
