+++ src/refactoring/components/admin/product/ProductManagementItemView.tsx
@@ -1,4 +1,4 @@
-import { Product } from "../../../types";
+import { Product } from "../../../../types";
 
 interface ProductManagementItemViewProps {
   product: Product;
