+++ src/refactoring/components/admin/product/index.ts
@@ -1,0 +1,1 @@
+export * from "./ProductManagementSection";
